+++ src/app/builder/linked/[linkedBuildId]/layout.tsx
@@ -2,6 +2,7 @@ import { Metadata, ResolvingMetadata } from 'next'
 
 import getLinkedBuild from '@/app/(actions)/builds/get-linked-build'
 import { PageHeader } from '@/app/(components)/page-header'
+import { NAV_ITEMS } from '@/app/(types)/navigation'
 import { isErrorResponse } from '@/app/(utils)/is-error-response'
 
 export async function generateMetadata(
@@ -67,7 +68,7 @@ export async function generateMetadata(
   }
 
   const title = `${linkedBuild.label} by ${createdByDisplayName}`
-  const description = `Link multiple builds together to create alternate versions of the same build.`
+  const description = NAV_ITEMS.linkedBuilds.description
 
   return {
     title,